Adds API code cascading
Most of the code is now in the API. Controller is just an epmty shell
that directs CRUD calls to the API controller.
The API handles this an know when to return a json error (direct API
call) or when it's called from a cousin controller. This part should
probably revisited to lift the ::API:: assumption in the restify helper.